p4i integration Integrate change 64304 from 3.0 to main. Description for change 64304:
: Change 64304 by vohi@vohi-hokkaido-nt on 2002/05/26 17:45:45
: 
: 	FIXED: QApplication
: 	Windows only - support mouse wheel on multihead setups with negative coordinates.
: 	
: 	qt-bugs/arc-18/36038
: 
: Affected files ...
: 
: ... //depot/qt/3.0/src/kernel/qapplication_win.cpp#51 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 64305]

+++ src/kernel/qapplication_win.cpp
@@ -3075,8 +3075,8 @@ bool QETWidget::translateWheelEvent( const MSG &msg )
 
     QPoint globalPos;
 
-    globalPos.rx() = LOWORD ( msg.lParam );
-    globalPos.ry() = HIWORD ( msg.lParam );
+    globalPos.rx() = (short)LOWORD ( msg.lParam );
+    globalPos.ry() = (short)HIWORD ( msg.lParam );
 
 
     // if there is a widget under the mouse and it is not shadowed
